# backend/swagger.yaml
openapi: 3.0.3
info:
  title: Intelligence Reports API
  description: |
    A secure API for managing intelligence reports with authentication and search capabilities.
    
    ## Authentication
    All endpoints except the health check require an API key to be provided in the `X-API-Key` header.
    
    ## Features
    - Create new intelligence reports
    - Retrieve specific reports by ID
    - Search and filter reports by tags and content
    - Secure API key authentication
    
  version: 1.0.0
  contact:
    name: API Support
  license:
    name: MIT

servers:
  - url: http://localhost:8000
    description: Development server
  - url: https://api.example.com
    description: Production server

security:
  - ApiKeyAuth: []

paths:
  /:
    get:
      summary: Health Check
      description: Check if the API is running
      operationId: healthCheck
      security: []
      responses:
        '200':
          description: API is running
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Intelligence Reports API is running"

  /report:
    post:
      summary: Create New Report
      description: Create a new intelligence report
      operationId: createReport
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateReport'
            example:
              title: "Security Threat Analysis"
              content: "Detailed analysis of potential security threats in the system..."
              tags: ["security", "threat-analysis", "urgent"]
      responses:
        '200':
          description: Report created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Report'
              example:
                id: "123e4567-e89b-12d3-a456-426614174000"
                title: "Security Threat Analysis"
                content: "Detailed analysis of potential security threats in the system..."
                tags: ["security", "threat-analysis", "urgent"]
                date: "2024-01-15T10:30:00.000Z"
        '400':
          description: Bad request - Missing or invalid request body
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                missing_body:
                  summary: Missing request body
                  value:
                    detail: "Request body is required"
                    status_code: 400
                invalid_json:
                  summary: Invalid JSON format
                  value:
                    detail: "Invalid JSON format in request body"
                    status_code: 400
                empty_body:
                  summary: Empty request body
                  value:
                    detail: "Request body cannot be empty"
                    status_code: 400
        '401':
          description: Unauthorized - Invalid or missing API key
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                missing_api_key:
                  summary: Missing API key
                  value:
                    detail: "API key required. Please provide X-API-Key header."
                invalid_api_key:
                  summary: Invalid API key
                  value:
                    detail: "Invalid API key."
        '422':
          description: Validation error - Invalid field values
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationErrorResponse'
              examples:
                missing_title:
                  summary: Missing title field
                  value:
                    detail: "Validation error: One or more fields are invalid"
                    errors:
                      - field: "title"
                        message: "Field required"
                        value: "{'content': 'Some content', 'tags': ['tag1', 'tag2']}"
                    status_code: 422
                empty_title:
                  summary: Empty title field
                  value:
                    detail: "Validation error: One or more fields are invalid"
                    errors:
                      - field: "title"
                        message: "String should have at least 1 character"
                        value: ""
                    status_code: 422
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                database_error:
                  summary: Database error
                  value:
                    detail: "Error during creating report: Database connection failed"
                    status_code: 500

  /report/{report_id}:
    get:
      summary: Get Report by ID
      description: Retrieve a specific intelligence report by its ID
      operationId: getReport
      parameters:
        - name: report_id
          in: path
          required: true
          description: Unique identifier of the report
          schema:
            type: string
            format: uuid
          example: "123e4567-e89b-12d3-a456-426614174000"
      responses:
        '200':
          description: Report found successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Report'
        '401':
          description: Unauthorized - Invalid or missing API key
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                missing_api_key:
                  summary: Missing API key
                  value:
                    detail: "API key required. Please provide X-API-Key header."
        '404':
          description: Report not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                report_not_found:
                  summary: Report not found
                  value:
                    detail: "Report with ID '123e4567-e89b-12d3-a456-426614174000' not found"
                    status_code: 404
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                database_error:
                  summary: Database error
                  value:
                    detail: "Error during retrieving report: Database connection failed"
                    status_code: 500

  /reports:
    get:
      summary: Get All Reports
      description: Retrieve all intelligence reports with optional filtering and search
      operationId: getReports
      parameters:
        - name: tag
          in: query
          required: false
          description: Filter reports by a specific tag
          schema:
            type: string
          example: "security"
        - name: search
          in: query
          required: false
          description: Search text in report title and content
          schema:
            type: string
          example: "threat analysis"
      responses:
        '200':
          description: Reports retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Report'
              example:
                - id: "123e4567-e89b-12d3-a456-426614174000"
                  title: "Security Threat Analysis"
                  content: "Detailed analysis of potential security threats..."
                  tags: ["security", "threat-analysis"]
                  date: "2024-01-15T10:30:00.000Z"
                - id: "987fcdeb-51a2-43d7-8f9e-123456789abc"
                  title: "Network Vulnerability Report"
                  content: "Assessment of network security vulnerabilities..."
                  tags: ["network", "vulnerability"]
                  date: "2024-01-14T15:45:00.000Z"
        '401':
          description: Unauthorized - Invalid or missing API key
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                missing_api_key:
                  summary: Missing API key
                  value:
                    detail: "API key required. Please provide X-API-Key header."
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                database_error:
                  summary: Database error
                  value:
                    detail: "Error during retrieving reports: Database connection failed"
                    status_code: 500

components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key for authentication

  schemas:
    Report:
      type: object
      required:
        - id
        - title
        - content
        - tags
        - date
      properties:
        id:
          type: string
          format: uuid
          description: Unique identifier for the report
          example: "123e4567-e89b-12d3-a456-426614174000"
        title:
          type: string
          description: Title of the intelligence report
          example: "Security Threat Analysis"
          maxLength: 200
        content:
          type: string
          description: Detailed content of the report
          example: "Detailed analysis of potential security threats in the system..."
        tags:
          type: array
          items:
            type: string
          description: List of tags associated with the report
          example: ["security", "threat-analysis", "urgent"]
          minItems: 1
        date:
          type: string
          format: date-time
          description: Date and time when the report was created
          example: "2024-01-15T10:30:00.000Z"

    CreateReport:
      type: object
      required:
        - title
        - content
        - tags
      properties:
        title:
          type: string
          description: Title of the intelligence report
          example: "Security Threat Analysis"
          maxLength: 200
        content:
          type: string
          description: Detailed content of the report
          example: "Detailed analysis of potential security threats in the system..."
        tags:
          type: array
          items:
            type: string
          description: List of tags associated with the report
          example: ["security", "threat-analysis", "urgent"]
          minItems: 1

    ErrorResponse:
      type: object
      required:
        - detail
        - status_code
      properties:
        detail:
          type: string
          description: Error message
          example: "Request body is required"
        status_code:
          type: integer
          description: HTTP status code
          example: 400

    ValidationErrorResponse:
      type: object
      required:
        - detail
        - errors
        - status_code
      properties:
        detail:
          type: string
          description: General validation error message
          example: "Validation error: One or more fields are invalid"
        errors:
          type: array
          items:
            type: object
            required:
              - field
              - message
            properties:
              field:
                type: string
                description: Field name that failed validation
                example: "title"
              message:
                type: string
                description: Specific validation error message
                example: "Field required"
              value:
                type: string
                description: The value that failed validation (optional)
                example: ""
          description: List of specific validation errors
        status_code:
          type: integer
          description: HTTP status code
          example: 422

tags:
  - name: Reports
    description: Operations related to intelligence reports
  - name: Health
    description: API health and status operations